Added interactivity to the SVG, contributed by GeorgeEdwards.
If it's loaded within a web-browser you can get the values at any point by
clicking and moving the mouse. This only seems to work if you open the SVG
file directly and not via the HTML page.

Put the plot point text back. It's in teeny tiny Eyestrain-O-Vision if you want to zoom in.